.github/workflows/scripts/Changelog-Generator/requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Deprecated==1.2.18
idna==3.10
iniconfig==2.0.0
loguru==0.7.3
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PyGithub==2.6.1
PyJWT==2.10.1
PyNaCl==1.5.0
pytest==8.3.5
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
